fix the size of the SUB_HEADER_SIZE-constant in Town_PatchSpec

+++ tools/patcher/townp.cpp
@@ -6461,7 +6461,7 @@ BYTE* Town_PatchSpec(const BYTE* minBuf, size_t minLen, const BYTE* celBuf, size
 	};
 
 	constexpr BYTE TRANS_COLOR = 128;
-	constexpr BYTE SUB_HEADER_SIZE = 14;
+	constexpr uint16_t SUB_HEADER_SIZE = 14;
 	constexpr int FRAME_WIDTH = TILE_WIDTH;
 	constexpr int FRAME_HEIGHT = 7 * TILE_HEIGHT;
